pyproject.toml
[project]
name = "lost-and-found"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "clip",
    "faiss-cpu>=1.13.0",
    "fastapi>=0.122.0",
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "python-multipart>=0.0.20",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "uvicorn>=0.38.0",
]

[tool.uv.sources]
clip = { git = "https://github.com/openai/CLIP.git" }
